+++ apps/wallet/src/background/trpc/adapter/types.ts
@@ -1,0 +1,22 @@
+import type {
+    TRPCClientOutgoingMessage,
+    TRPCErrorResponse,
+    TRPCRequest,
+    TRPCResultMessage,
+} from '@trpc/server/rpc';
+
+export type TRPCExtensionRequest = {
+    trpc: TRPCRequest | TRPCClientOutgoingMessage;
+};
+
+export type TRPCExtensionSuccessResponse = {
+    trpc: TRPCResultMessage<unknown>;
+};
+
+export type TRPCExtensionErrorResponse = {
+    trpc: TRPCErrorResponse;
+};
+
+export type TRPCExtensionResponse =
+    | TRPCExtensionSuccessResponse
+    | TRPCExtensionErrorResponse;
